GnollHackW starts again.

+++ win/win32/mhstatus.h
@@ -64,7 +64,7 @@ typedef struct mswin_status_field {
 } mswin_status_field;
 
 typedef struct mswin_condition_field {
-    int mask;
+    unsigned long mask;
     const char * name;
     int bit_position;
 } mswin_condition_field;
